requirements.txt
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
click==7.1.2
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.20
pathspec==0.8.0
pipenv==2020.6.2
protobuf==3.12.4
pycodestyle==2.6.0
pylint==2.5.2
PyMySQL==0.10.0
python-decouple==3.3
regex==2020.5.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.27
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
